requirements.txt
SolidPython
